pyproject.toml
[project]
name = "image-chatbot"
version = "0.1.0"
description = "A Flask-based image chatbot for plant disease detection"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "Flask>=3.0.0",
    "flask-cors>=6.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
